[REFACTOR] [core] - Rename test-utils directory.
