+++ packages/codistica-core/src/__tests__/modules/json-utils/internals/parse.test.js
@@ -1,6 +1,6 @@
 import {assert} from 'chai';
 import {parse} from '../../../../modules/json-utils/internals/parse.js';
-import {getCircularObject} from '../../../test-utils/get-circular-object.js';
+import {getCircularObject} from '../../../__utils__/get-circular-object.js';
 
 /** @see module:modules/json-utils/parse */
 function parseTest() {
