+++ packages/codistica-core/src/__tests__/modules/json-utils/internals/stringify.test.js
@@ -1,6 +1,6 @@
 import {assert} from 'chai';
 import {stringify} from '../../../../modules/json-utils/internals/stringify.js';
-import {getCircularObject} from '../../../test-utils/get-circular-object.js';
+import {getCircularObject} from '../../../__utils__/get-circular-object.js';
 
 /** @see module:modules/json-utils/stringify */
 function stringifyTest() {
