+++ packages/codistica-core/src/__tests__/modules/object-utils/internals/parse-circular.test.js
@@ -1,6 +1,6 @@
 import {assert} from 'chai';
 import {parseCircular} from '../../../../modules/object-utils/internals/parse-circular.js';
-import {getCircularObject} from '../../../test-utils/get-circular-object.js';
+import {getCircularObject} from '../../../__utils__/get-circular-object.js';
 
 /** @see module:codistica-core/modules/object-utils/parse-circular */
 function parseCircularTest() {
